✨ Feat: 유저 프로필 등록, 수정 구현

+++ backend/sleepony/src/main/java/com/c208/sleephony/domain/user/entity/User.java
@@ -2,6 +2,7 @@ package com.c208.sleephony.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -29,10 +30,10 @@ public class User {
     private String nickname;
 
     @Column(name = "height", precision = 4, scale = 1)
-    private double height;
+    private BigDecimal height;
 
     @Column(name = "weight", precision = 4, scale = 1)
-    private double weight;
+    private BigDecimal weight;
 
     @Column(name = "birth_date")
     private LocalDate birthDate;
